requirements.txt
Flask
python-dotenv
llama-index
llama-index-embeddings-azure-openai
llama-index-readers-web
llama-cpp-python
torch==2.0.1
transformers
langcheck
html2text
gunicorn